call getLastKnownLocation on locationClient inside LocationService.onConnected (LocationService.java) in order to provide a startup location for the listeners.
This should fix the missing location information at startup.

http://stackoverflow.com/questions/11183064/wifi-provider-onlocationchanged-is-never-called/14012799#14012799

" 0 down vote
	

I had the very same problem and in turns out that onLocationChanged is
not called if there is getLastKnownLocation which is the same as
"current". That means that you will not get new location because there
is nothing new to report... Solution is to use the last known location
even it is very old (and continue to listen for updates if you need more
precise/new location (if such occurres)).
"